Added dtype parameter in dpnp.sum() function call for avoid copying.

+++ dpbench/benchmarks/l2_norm/l2_norm_dpnp.py
@@ -7,5 +7,5 @@ import dpnp as np
 
 def l2_norm(a, d):
     sq = np.square(a)
-    sum = sq.sum(axis=1)
+    sum = sq.sum(axis=1, dtype=sq.dtype)
     d[:] = np.sqrt(sum)
